ci: use port 80 for api
