+++ libs/backend/infra/adapter/graphql/src/index.ts
@@ -1,3 +1,4 @@
 export * from './auth.middleware'
 export * from './context'
 export * from './cors.middleware'
+export * from './proxy.middleware'
